refactor: 변수명 변경타, 타입 변경 로직 변경

+++ src/main/java/org/springeel/oneclickrecipe/domain/recipe/dto/service/response/RecipeAllReadResponseDto.java
@@ -11,7 +11,7 @@ public record RecipeAllReadResponseDto(
     String imageUrl,
     String writer,
     Boolean isLiked,
-    Long likeCount
+    Integer likeCount
 ) {
 
 }
